# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: STRESSNES luxury fashion ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - in: query
          name: productId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReviewList"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a product review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a review by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update review status (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Review updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── Wishlist ─────────────────────────────────────────────────
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get the current user's wishlist
      responses:
        "200":
          description: Wishlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"

    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add a product to the wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Item added
  /wishlist/{itemId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove a product from the wishlist
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item removed

  # ── Coupons ──────────────────────────────────────────────────
...
components:
  schemas:
...
    ProductImage:
      type: object
      required: [id, url]
      properties:
        id:
          type: string
        url:
          type: string
        altText:
          type: ["string", "null"]
        isPrimary:
          type: boolean
        sortOrder:
          type: integer
    ProductVariant:
      type: object
      required: [id, sku, isActive]
      properties:
        id:
          type: string
        sku:
          type: string
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        material:
          type: ["string", "null"]
        priceOverride:
          type: ["number", "null"]
        isActive:
          type: boolean
        stockQty:
          type: integer
    ProductCategoryRef:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
    Product:
      type: object
      required: [id, title, slug, sku, price, status]
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        sku:
          type: string
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        status:
          type: string
        featured:
          type: boolean
        published:
          type: boolean
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        category:
          oneOf:
            - $ref: "#/components/schemas/ProductCategoryRef"
            - type: "null"
        collection:
          oneOf:
            - $ref: "#/components/schemas/ProductCategoryRef"
            - type: "null"
        createdAt:
          type: string
          format: date-time
...
    Review:
      type: object
      required: [id, productId, rating, status, createdAt]
      properties:
        id:
          type: string
        productId:
          type: string
        userId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    PaginatedReviewList:
      type: object
      required: [data, total, page, pageSize, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
    ReviewInput:
      type: object
      required: [productId, rating]
      properties:
        productId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
    ReviewUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]

    # ── Wishlist ───────────────────────────────────────────────
    WishlistItem:
      type: object
      required: [id, productId, addedAt]
      properties:
        id:
          type: string
        productId:
          type: string
        addedAt:
          type: string
          format: date-time
        product:
          $ref: "#/components/schemas/Product"
    WishlistInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: string

    # ── Coupons ────────────────────────────────────────────────
...